+++ source/world_builder/features/oceanic_plate_models/composition/tian2019_water_content.cc
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2018 - 2021 by the authors of the World Builder code.
+  Copyright (C) 2018 - 2024 by the authors of the World Builder code.
 
   This file is part of the World Builder.
 
@@ -17,7 +17,7 @@
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "world_builder/features/oceanic_plate_models/composition/water_content.h"
+#include "world_builder/features/oceanic_plate_models/composition/tian2019_water_content.h"
 
 #include "world_builder/kd_tree.h"
 #include "world_builder/nan.h"
@@ -28,6 +28,7 @@
 #include "world_builder/types/unsigned_int.h"
 #include "world_builder/types/value_at_points.h"
 #include "world_builder/utilities.h"
+#include "world_builder/world.h"
 
 
 namespace WorldBuilder
@@ -41,7 +42,7 @@ namespace WorldBuilder
     {
       namespace Composition
       {
-        WaterContent::WaterContent(WorldBuilder::World *world_)
+        TianWaterContent::TianWaterContent(WorldBuilder::World *world_)
           :
           min_depth(NaN::DSNAN),
           max_depth(NaN::DSNAN),
@@ -51,16 +52,23 @@ namespace WorldBuilder
           this->name = "water content";
         }
 
-        WaterContent::~WaterContent()
+        TianWaterContent::~TianWaterContent()
           = default;
 
         void
-        WaterContent::declare_entries(Parameters &prm, const std::string & /*unused*/)
+        TianWaterContent::declare_entries(Parameters &prm, const std::string & /*unused*/)
         {
           // Document plugin and require entries if needed.
           // Add compositions to the required parameters.
           prm.declare_entry("", Types::Object({"compositions"}),
-                            "WaterContent compositional model. Sets water content as a compositional field.");
+                            "TianWaterContent compositional model. Sets bound water content as a compositional field. The returned "
+                            "water content is based on the the temperature and pressure at a point within the world. Currently, "
+                            "the bound water content can be determined for four different lithologies: 'sediment', mid-ocean "
+                            "ridge basalt ('MORB'), 'gabbro', and 'peridotite', using parameterized phase diagrams from Tian et al., 2019 "
+                            "(https://doi.org/10.1029/2019GC008488). The pressure is lithostatic, calculated with a constant user defined "
+                            "density, and is limited by a user defined cutoff pressure (in GPa) for each lithology. This is required because the "
+                            "parameterization breaks down at large pressures. Recommended cutoff pressures are 10 GPa is used for 'peridotite', "
+                            "26 GPa is used for 'gabbro', 16 GPa is used for 'MORB', and 1 GPa is used for 'sediment'.");
 
           // Declare entries of this plugin
           prm.declare_entry("min depth", Types::OneOf(Types::Double(0),Types::Array(Types::ValueAtPoints(0., 2.))),
@@ -70,22 +78,26 @@ namespace WorldBuilder
           prm.declare_entry("compositions", Types::Array(Types::UnsignedInt(),0),
                             "A list with the labels of the composition which are present there.");
           prm.declare_entry("density", Types::Double(3000.0),
-                            "The reference density used to estimate the lithostatic pressure for calculating "
+                            "The reference density used for determining the lithostatic pressure for calculating "
                             "the bound water content.");
           prm.declare_entry("lithology",  Types::String("peridotite"),
-                            "The lithology used to determine which polynomials to use for calculating the water content.");
+                            "The lithology used to determine which polynomials to use for calculating the water content. Valid options are: "
+                            "'sediment', 'MORB', 'gabbro', and 'peridotite'.");
           prm.declare_entry("initial water content", Types::Double(5),
-                            "The value of the initial water content (in wt%) for the lithology at the trench. This is essentially the "
+                            "The value of the initial water content (in wt%) for the lithology at the trench. This represents the "
                             "max value applied to this lithology.");
+          prm.declare_entry("cutoff pressure", Types::Double(10),
+                            "The upper bound for the pressure, in GPa, for the specified lithology in the Tian parameterization. This is necessary because "
+                            "the parameterization breaks down for high pressures. It is recommended that 10 GPa is used for 'peridotite', 26 GPa is used for "
+                            "'gabbro', 16 GPa is used for 'MORB', and 1 GPa is used for 'sediment'.");
           prm.declare_entry("operation", Types::String("replace", std::vector<std::string> {"replace", "replace defined only", "add", "subtract"}),
                             "Whether the value should replace any value previously defined at this location (replace) or "
                             "add the value to the previously define value. Replacing implies that all compositions not "
                             "explicitly defined are set to zero. To only replace the defined compositions use the replace only defined option.");
-
         }
 
         void
-        WaterContent::parse_entries(Parameters &prm, const std::vector<Point<2>> &coordinates)
+        TianWaterContent::parse_entries(Parameters &prm, const std::vector<Point<2>> &coordinates)
         {
           min_depth_surface = Objects::Surface(prm.get("min depth",coordinates));
           min_depth = min_depth_surface.minimum;
@@ -95,16 +107,24 @@ namespace WorldBuilder
           compositions = prm.get_vector<unsigned int>("compositions");
           max_water_content = prm.get<double>("initial water content");
           operation = string_operations_to_enum(prm.get<std::string>("operation"));
-          lithology_str = prm.get<std::string>("lithology");
+          cutoff_pressure = prm.get<double>("cutoff pressure");
+          std::string lithology_string = prm.get<std::string>("lithology");
+
+          if (lithology_string=="peridotite")
+            lithology_type = peridotite;
+          else if (lithology_string=="gabbro")
+            lithology_type = gabbro;
+          else if (lithology_string=="MORB")
+            lithology_type = MORB;
+          else if (lithology_string=="sediment")
+            lithology_type = sediment;
         }
 
 
         double
-        WaterContent::calculate_water_content(double pressure,
-                                              double temperature,
-                                              std::string lithology_string) const
+        TianWaterContent::calculate_water_content(double pressure,
+                                                  double temperature) const
         {
-          pressure = pressure <= 0.5 ? 0.5 : pressure;
           double ln_LR_value = 0;
           double ln_c_sat_value = 0;
           double Td_value = 0;
@@ -112,77 +132,25 @@ namespace WorldBuilder
           std::vector<double> c_sat_polynomial_coeffs;
           std::vector<double> Td_polynomial_coeffs;
 
-          enum LithologyName
-          {
-            peridotite,
-            gabbro,
-            MORB,
-            sediment
-          };
-          LithologyName lithology = peridotite;
-
-          if (lithology_string=="peridotite")
-            lithology = peridotite;
-          else if (lithology_string=="gabbro")
-            lithology = gabbro;
-          else if (lithology_string=="MORB")
-            lithology = MORB;
-          else if (lithology_string=="sediment")
-            lithology = sediment;
-
-          if (lithology == peridotite)
-            {
-              LR_polynomial_coeffs = LR_poly_peridotite;
-              c_sat_polynomial_coeffs = c_sat_poly_peridotite;
-              Td_polynomial_coeffs = Td_poly_peridotite;
-              pressure = pressure > pressure_cutoffs[0] ? pressure_cutoffs[0] : pressure;
-            }
-
-          if (lithology == gabbro)
-            {
-              LR_polynomial_coeffs = LR_poly_gabbro;
-              c_sat_polynomial_coeffs = c_sat_poly_gabbro;
-              Td_polynomial_coeffs = Td_poly_gabbro;
-              pressure = pressure > pressure_cutoffs[1] ? pressure_cutoffs[1] : pressure;
-            }
-
-          if (lithology == MORB)
-            {
-              LR_polynomial_coeffs = LR_poly_MORB;
-              c_sat_polynomial_coeffs = c_sat_poly_MORB;
-              Td_polynomial_coeffs = Td_poly_MORB;
-              pressure = pressure > pressure_cutoffs[2] ? pressure_cutoffs[2] : pressure;
-            }
-
-          if (lithology == sediment)
-            {
-              LR_polynomial_coeffs = LR_poly_sediment;
-              c_sat_polynomial_coeffs = c_sat_poly_sediment;
-              Td_polynomial_coeffs = Td_poly_sediment;
-              pressure = pressure > pressure_cutoffs[3] ? pressure_cutoffs[3] : pressure;
-            }
-
-          double inv_pressure = 1/pressure;
-
           // Calculate the c_sat value from Tian et al., 2019
-          if (lithology == sediment)
+          if (lithology_type == sediment)
             {
-              for (unsigned int c_sat_index = 0; c_sat_index < c_sat_polynomial_coeffs.size(); ++c_sat_index)
-                ln_c_sat_value += c_sat_polynomial_coeffs[c_sat_index] * (std::pow(std::log10(pressure), c_sat_polynomial_coeffs.size() - 1 - c_sat_index));
+              for (unsigned int c_sat_index = 0; c_sat_index < c_sat_poly[lithology_type].size(); ++c_sat_index)
+                ln_c_sat_value += c_sat_poly[lithology_type][c_sat_index] * (std::pow(std::log10(pressure), c_sat_poly[lithology_type].size() - 1 - c_sat_index));
             }
           else
             {
-              for (unsigned int c_sat_index = 0; c_sat_index < c_sat_polynomial_coeffs.size(); ++c_sat_index)
-                ln_c_sat_value += c_sat_polynomial_coeffs[c_sat_index] * (std::pow(pressure, c_sat_polynomial_coeffs.size() - 1 - c_sat_index));
+              for (unsigned int c_sat_index = 0; c_sat_index < c_sat_poly[lithology_type].size(); ++c_sat_index)
+                ln_c_sat_value += c_sat_poly[lithology_type][c_sat_index] * (std::pow(pressure, c_sat_poly[lithology_type].size() - 1 - c_sat_index));
             }
 
           // Calculate the LR value from Tian et al., 2019
-          for (unsigned int LR_coeff_index = 0; LR_coeff_index < LR_polynomial_coeffs.size(); ++LR_coeff_index)
-            ln_LR_value += LR_polynomial_coeffs[LR_coeff_index] * (std::pow(inv_pressure, LR_polynomial_coeffs.size() - 1 - LR_coeff_index));
+          for (unsigned int LR_coeff_index = 0; LR_coeff_index < LR_poly[lithology_type].size(); ++LR_coeff_index)
+            ln_LR_value += LR_poly[lithology_type][LR_coeff_index] * (std::pow(1/pressure, LR_poly[lithology_type].size() - 1 - LR_coeff_index));
 
           // Calculate the Td value from Tian et al., 2019
-          for (unsigned int Td_coeff_index = 0; Td_coeff_index < Td_polynomial_coeffs.size(); ++Td_coeff_index)
-            Td_value += Td_polynomial_coeffs[Td_coeff_index] * (std::pow(pressure, Td_polynomial_coeffs.size() - 1 - Td_coeff_index));
+          for (unsigned int Td_coeff_index = 0; Td_coeff_index < Td_poly[lithology_type].size(); ++Td_coeff_index)
+            Td_value += Td_poly[lithology_type][Td_coeff_index] * (std::pow(pressure, Td_poly[lithology_type].size() - 1 - Td_coeff_index));
 
           double partition_coeff = std::exp(ln_c_sat_value) * std::exp(std::exp(ln_LR_value) * (1/temperature - 1/Td_value));
           return partition_coeff;
@@ -190,13 +158,13 @@ namespace WorldBuilder
 
 
         double
-        WaterContent::get_composition(const Point<3> &position_in_cartesian_coordinates,
-                                      const Objects::NaturalCoordinate &position_in_natural_coordinates,
-                                      const double depth,
-                                      const unsigned int composition_number,
-                                      double composition,
-                                      const double  /*feature_min_depth*/,
-                                      const double  /*feature_max_depth*/) const
+        TianWaterContent::get_composition(const Point<3> &position_in_cartesian_coordinates,
+                                          const Objects::NaturalCoordinate &position_in_natural_coordinates,
+                                          const double depth,
+                                          const unsigned int composition_number,
+                                          double composition,
+                                          const double  /*feature_min_depth*/,
+                                          const double  /*feature_max_depth*/) const
         {
           if (depth <= max_depth && depth >= min_depth)
             {
@@ -204,11 +172,12 @@ namespace WorldBuilder
               const double max_depth_local = max_depth_surface.constant_value ? max_depth : max_depth_surface.local_value(position_in_natural_coordinates.get_surface_point()).interpolated_value;
               if (depth <= max_depth_local &&  depth >= min_depth_local)
                 {
-                  double lithostatic_pressure = density * 9.81 * depth / 1e9; // GPa
-                  double slab_temperature = this->world->properties(position_in_cartesian_coordinates.get_array(), depth, {{{1,0,0}}})[0];
+                  // The polynomials break down for pressures less than 0.5 GPa, and for pressures above a user defined cutoff pressure
+                  // ensure that the pressure is never below 0.5 GPa
+                  const double lithostatic_pressure = std::max(0.5, std::min(density * 9.81 * depth / 1e9, cutoff_pressure)); // GPa
+                  const double slab_temperature = world->properties(position_in_cartesian_coordinates.get_array(), depth, {{{1,0,0}}})[0];
                   double partition_coefficient = calculate_water_content(lithostatic_pressure,
-                                                                         slab_temperature,
-                                                                         lithology_str);
+                                                                         slab_temperature);
 
                   partition_coefficient = std::min(max_water_content, partition_coefficient);
 
@@ -228,7 +197,7 @@ namespace WorldBuilder
             }
           return composition;
         }
-        WB_REGISTER_FEATURE_OCEANIC_PLATE_COMPOSITION_MODEL(WaterContent, water content)
+        WB_REGISTER_FEATURE_OCEANIC_PLATE_COMPOSITION_MODEL(TianWaterContent, tian water content)
       } // namespace Composition
     } // namespace OceanicPlateModels
   } // namespace Features
