+++ tests/gwb-dat/water_content_subducting_plate/screen-output.log
@@ -1,0 +1,24 @@
+# x z d T c0 tag 
+50e3 0e3 1e3 312.405 3  0
+50e3 0e3 4e3 430.229 1  0
+50e3 0e3 8e3 584.338 0.5  0
+50e3 0e3 11e3 695.831 2  0
+210e3 0e3 55e3 1322.25 0  1
+210e3 0e3 57.5e3 1203.62 0.372427  1
+210e3 0e3 60e3 1034.33 0.510104  1
+210e3 0e3 62.5e3 849.688 0.689769  1
+210e3 0e3 65e3 697.68 1  1
+210e3 0e3 67.5e3 614.688 1  1
+210e3 0e3 70e3 616.639 0.5  1
+210e3 0e3 72.5e3 652.908 0.5  1
+210e3 0e3 75e3 687.783 0.5  1
+210e3 0e3 77.5e3 721.222 2  1
+210e3 0e3 80e3 753.193 2  1
+210e3 0e3 82.5e3 783.675 2  1
+210e3 0e3 85e3 812.658 2  1
+210e3 0e3 87.5e3 839.142 2  1
+210e3 0e3 90e3 863.691 2  1
+210e3 0e3 92.5e3 886.838 1.91775  1
+210e3 0e3 95e3 908.585 1.30946  1
+210e3 0e3 97.5e3 928.938 0.908444  1
+210e3 0e3 100e3 947.908 0.640881  1
